refactor(condo): DOMA-7815 added comments for readability
